requirements-extra.txt
dj-database-url
psycopg2-binary
django-storages[boto3]
